Configure audio and video configurations

+++ app/src/main/java/xyz/juncat/media/record/screen/config/VideoConfig.kt
@@ -3,7 +3,7 @@ package xyz.juncat.media.record.screen.config
 data class VideoConfig(
     val width: Int,
     val height: Int,
-    val bitrate: Long,
+    val bitrate: Int,
     val fps: Int,
     val iFrameInterval: Int,
     val fpsMode: FPSMode,
